+++ HISTORY.md
@@ -1,6 +1,10 @@
 History
 =======
 
+## 0.0.3
+
+* Add `baseMeta` option to `error`, `uncaught` handlers.
+
 ## 0.0.2
 
 * Add `error`, `uncaught` handlers.